move sun&avg composite index tests to new file
