+++ packages/firestore/main.tf
@@ -11,15 +11,36 @@ resource "google_firestore_index" "default-db-index" {
   dynamic "fields" {
     for_each = distinct(flatten([for k, v in local.indexes : [
       for i in each.value : {
-        field_path = i.field_path
-        order      = i.order
+        field_path   = i.field_path
+        order        = can(i.order) ? i.order : null
+        array_config = can(i.array_config) ? i.array_config : null
     }]]))
     content {
-      field_path = lookup(fields.value, "field_path", null)
-      order      = lookup(fields.value, "order", null)
+      field_path   = fields.value.field_path
+      order        = fields.value.order
+      array_config = fields.value.array_config
     }
   }
+}
 
+resource "google_firestore_index" "default-db-collection-group-index" {
+  collection  = "composite-index-test-collection"
+  query_scope = "COLLECTION_GROUP"
+
+  for_each = local.collection_group_indexes
+  dynamic "fields" {
+    for_each = distinct(flatten([for k, v in local.indexes : [
+      for i in each.value : {
+        field_path   = i.field_path
+        order        = can(i.order) ? i.order : null
+        array_config = can(i.array_config) ? i.array_config : null
+    }]]))
+    content {
+      field_path   = fields.value.field_path
+      order        = fields.value.order
+      array_config = fields.value.array_config
+    }
+  }
 }
 
 resource "google_firestore_index" "named-db-index" {
@@ -30,12 +51,35 @@ resource "google_firestore_index" "named-db-index" {
   dynamic "fields" {
     for_each = distinct(flatten([for k, v in local.indexes : [
       for i in each.value : {
-        field_path = i.field_path
-        order      = i.order
+        field_path   = i.field_path
+        order        = can(i.order) ? i.order : null
+        array_config = can(i.array_config) ? i.array_config : null
     }]]))
     content {
-      field_path = lookup(fields.value, "field_path", null)
-      order      = lookup(fields.value, "order", null)
+      field_path   = fields.value.field_path
+      order        = fields.value.order
+      array_config = fields.value.array_config
+    }
+  }
+}
+
+resource "google_firestore_index" "named-db-collection-group-index" {
+  collection  = "composite-index-test-collection"
+  database    = "test-db"
+  query_scope = "COLLECTION_GROUP"
+
+  for_each = local.collection_group_indexes
+  dynamic "fields" {
+    for_each = distinct(flatten([for k, v in local.indexes : [
+      for i in each.value : {
+        field_path   = i.field_path
+        order        = can(i.order) ? i.order : null
+        array_config = can(i.array_config) ? i.array_config : null
+    }]]))
+    content {
+      field_path   = fields.value.field_path
+      order        = fields.value.order
+      array_config = fields.value.array_config
     }
   }
 }
