+++ packages/firestore/test/integration/api/aggregation.test.ts
@@ -1385,156 +1385,4 @@ apiDescribe('Aggregation queries - sum / average', persistence => {
       expect(snapshot.data().countOfDocs).to.equal(4);
     });
   });
-
-  // Only run tests that require indexes against the emulator, because we don't
-  // have a way to dynamically create the indexes when running the tests.
-  (USE_EMULATOR ? apiDescribe : apiDescribe.skip)(
-    'queries requiring indexes',
-    () => {
-      it('aggregate query supports collection groups - multi-aggregate', () => {
-        return withTestDb(persistence, async db => {
-          const collectionGroupId = doc(
-            collection(db, 'aggregateQueryTest')
-          ).id;
-          const docPaths = [
-            `${collectionGroupId}/cg-doc1`,
-            `abc/123/${collectionGroupId}/cg-doc2`,
-            `zzz${collectionGroupId}/cg-doc3`,
-            `abc/123/zzz${collectionGroupId}/cg-doc4`,
-            `abc/123/zzz/${collectionGroupId}`
-          ];
-          const batch = writeBatch(db);
-          for (const docPath of docPaths) {
-            batch.set(doc(db, docPath), { x: 2 });
-          }
-          await batch.commit();
-          const snapshot = await getAggregateFromServer(
-            collectionGroup(db, collectionGroupId),
-            {
-              count: count(),
-              sum: sum('x'),
-              avg: average('x')
-            }
-          );
-          expect(snapshot.data().count).to.equal(2);
-          expect(snapshot.data().sum).to.equal(4);
-          expect(snapshot.data().avg).to.equal(2);
-        });
-      });
-
-      it('performs aggregations on documents with all aggregated fields using getAggregationFromServer', () => {
-        const testDocs = {
-          a: { author: 'authorA', title: 'titleA', pages: 100, year: 1980 },
-          b: { author: 'authorB', title: 'titleB', pages: 50, year: 2020 },
-          c: { author: 'authorC', title: 'titleC', pages: 150, year: 2021 },
-          d: { author: 'authorD', title: 'titleD', pages: 50 }
-        };
-        return withTestCollection(persistence, testDocs, async coll => {
-          const snapshot = await getAggregateFromServer(coll, {
-            totalPages: sum('pages'),
-            averagePages: average('pages'),
-            averageYear: average('year'),
-            count: count()
-          });
-          expect(snapshot.data().totalPages).to.equal(300);
-          expect(snapshot.data().averagePages).to.equal(100);
-          expect(snapshot.data().averageYear).to.equal(2007);
-          expect(snapshot.data().count).to.equal(3);
-        });
-      });
-
-      it('performs aggregates on multiple fields where one aggregate could cause short-circuit due to NaN using getAggregationFromServer', () => {
-        const testDocs = {
-          a: {
-            author: 'authorA',
-            title: 'titleA',
-            pages: 100,
-            year: 1980,
-            rating: 5
-          },
-          b: {
-            author: 'authorB',
-            title: 'titleB',
-            pages: 50,
-            year: 2020,
-            rating: 4
-          },
-          c: {
-            author: 'authorC',
-            title: 'titleC',
-            pages: 100,
-            year: 1980,
-            rating: Number.NaN
-          },
-          d: {
-            author: 'authorD',
-            title: 'titleD',
-            pages: 50,
-            year: 2020,
-            rating: 0
-          }
-        };
-        return withTestCollection(persistence, testDocs, async coll => {
-          const snapshot = await getAggregateFromServer(coll, {
-            totalRating: sum('rating'),
-            totalPages: sum('pages'),
-            averageYear: average('year')
-          });
-          expect(snapshot.data().totalRating).to.be.NaN;
-          expect(snapshot.data().totalPages).to.equal(300);
-          expect(snapshot.data().averageYear).to.equal(2000);
-        });
-      });
-
-      it('performs aggregates when using `array-contains-any` operator getAggregationFromServer', () => {
-        const testDocs = {
-          a: {
-            author: 'authorA',
-            title: 'titleA',
-            pages: 100,
-            year: 1980,
-            rating: [5, 1000]
-          },
-          b: {
-            author: 'authorB',
-            title: 'titleB',
-            pages: 50,
-            year: 2020,
-            rating: [4]
-          },
-          c: {
-            author: 'authorC',
-            title: 'titleC',
-            pages: 100,
-            year: 1980,
-            rating: [2222, 3]
-          },
-          d: {
-            author: 'authorD',
-            title: 'titleD',
-            pages: 50,
-            year: 2020,
-            rating: [0]
-          }
-        };
-        return withTestCollection(persistence, testDocs, async coll => {
-          const snapshot = await getAggregateFromServer(
-            query(coll, where('rating', 'array-contains-any', [5, 3])),
-            {
-              totalRating: sum('rating'),
-              averageRating: average('rating'),
-              totalPages: sum('pages'),
-              averagePages: average('pages'),
-              countOfDocs: count()
-            }
-          );
-          expect(snapshot.data().totalRating).to.equal(0);
-          expect(snapshot.data().averageRating).to.be.null;
-          expect(snapshot.data().totalPages).to.equal(200);
-          expect(snapshot.data().averagePages).to.equal(100);
-          expect(snapshot.data().countOfDocs).to.equal(2);
-        });
-      });
-    }
-  );
 });
